third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/029-expected.txt

This is a testharness.js-based test.
FAIL  scheduler: javascript: URL in HREF assert_array_equals: property 2, expected "beforeunload event" but got "inline script #2"

assert_array_equals: property 3, expected "beforeunload event" but got "JS URL"
Harness: the test ran to completion.
